+++ src/dra_psp/199E8.c
@@ -1147,7 +1147,7 @@ void AnimateNowLoading(NowLoadingModel* self, s16 x, s16 y, bool isDone) {
     s16 posX;
     s16 posY;
     s16 angle;
-    s16 verticalWave;
+    s32 verticalWave;
     s32 horizontalWave;
     s32 baseY;
     s16 sx;
